working time gestion
ajout de'un systeme de gestion des service et des trips en fonction du temps (actives ou pas )  modification de la maniere de donner un poid au edges

formatting the code + removing unused import

+++ MyWay/src/ma/myway/graph/Edge.java
@@ -24,19 +24,19 @@ import java.util.Date;
  * 
  * @see Node
  */
-public class Edge implements Serializable{
+public class Edge implements Serializable {
 	/**
 	 * 
 	 */
 	private static final long serialVersionUID = 160843439856388648L;
 	private Node src, dest;
-	private double weight; // type a discuter
+	private long weight; // type a discuter
 	private String trip_id;
 	private int transfer_type;
-	private boolean isTransfert ;
+	private boolean isTransfert;
 	private boolean isActive;
 
-	public Edge(Node src, Node dest, double weight, String trip_id) {
+	public Edge(Node src, Node dest, long weight, String trip_id) {
 		this.src = src;
 		this.dest = dest;
 		this.weight = weight;
@@ -44,28 +44,30 @@ public class Edge implements Serializable{
 		this.isTransfert = false;
 	}
 
-	public Edge(Node src, Node dest, double weight, String trip_id, int transfer_type) {
+	public Edge(Node src, Node dest, long weight, String trip_id, int transfer_type) {
 		this.src = src;
 		this.dest = dest;
 		this.weight = weight;
 		this.trip_id = trip_id;
 		this.transfer_type = transfer_type;
 		this.isTransfert = true;
-		//this.isActive = activity();
+		// this.isActive = activity();
 	}
-	
+
 	/**
 	 * @return boolean
 	 */
 	public boolean isTransfert() {
 		return isTransfert;
 	}
+
 	/**
 	 * @return int
 	 */
 	public int getTransfer_type() {
 		return transfer_type;
 	}
+
 	/**
 	 * @return Node
 	 */
@@ -83,14 +85,14 @@ public class Edge implements Serializable{
 	/**
 	 * @return double
 	 */
-	public double getWeight() {
+	public long getWeight() {
 		return this.weight;
 	}
 
 	/**
 	 * @param weight
 	 */
-	public void setWeight(double weight) {
+	public void setWeight(long weight) {
 		this.weight = weight;
 	}
 
@@ -107,27 +109,20 @@ public class Edge implements Serializable{
 
 	@Override
 	public String toString() {
-		return "Edge [src=" + src + ", dest=" + dest.getStop().getStop_id() + ", weight=" + weight + ", trip_id=" + trip_id
-				+ ", transfer_type=" + transfer_type + ", isTransfert=" + isTransfert + "]";
+		return "Edge [src=" + src + ", dest=" + dest.getStop().getStop_id() + ", weight=" + weight + ", trip_id="
+				+ trip_id + ", transfer_type=" + transfer_type + ", isTransfert=" + isTransfert + "]";
 	}
-	
-	
 
-	public int getTransferType(){
+	public int getTransferType() {
 		throw new UnsupportedOperationException();
 	}
 
-	public LocalDate getNextStopTime(){
+	public LocalDate getNextStopTime() {
 		throw new UnsupportedOperationException();
 	}
-	
-	
-	
-	private static boolean activity (Date dateDep, Graph g) {
-		
-		
-		
-		
+
+	private static boolean activity(Date dateDep, Graph g) {
+
 		return false;
 	}
 }
